+++ GomokuAi/constants.h
@@ -4,4 +4,4 @@
 #define AI_DEBUG false
 
 extern int ScoreByNumArount[9];
-extern const uint32_t MaxNodeNum; //最大允许的节点数量（避免内存占用过多）
+extern const uint64_t MaxNodeNum; //最大允许的节点数量（避免内存占用过多）
